requirements.txt
Django==4.2
djangorestframework>=3.14.0
python-dotenv>=1.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2024.1
django-filter>=23.5
Pillow>=10.0.0
reportlab>=4.0.0
